requirements.txt
customtkinter
Pillow
pygame
